feat: add component tableOfSubject
